Add pre calulated lobanov formants

+++ src/plots/vowel-plot.R
@@ -3,30 +3,30 @@ list.of.packages <- c("ggplot2", "showtext", "reshape2", "plyr", "stringr")
 new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
 if(length(new.packages)) install.packages(new.packages)
 for (package in list.of.packages) {
-	library(package, character.only=TRUE)
+  library(package, character.only=TRUE)
 }
 
 rad <- function(a) a / 180 * pi
 
 lobanov <- function(df, f1="f1", f2="f2", vowel="vowel", group=c(), reduce=TRUE) {
   ddply(df, group, function(df.grp) {
-	  f1.grp <- df.grp[,f1]
-	  f2.grp <- df.grp[,f2]
-	  mn.f1.grp <- mean(f1.grp, na.rm=TRUE)
-	  mn.f2.grp <- mean(f2.grp, na.rm=TRUE)
-	  sd.f1.grp <- sd(f1.grp, na.rm=TRUE)
-	  sd.f2.grp <- sd(f2.grp, na.rm=TRUE)
-	  ddply(df.grp, vowel, function(df.vwl) {
-		  f1.vwl <- df.vwl[,f1]
-		  f2.vwl <- df.vwl[,f2]
-		  f1.vwl.nrm <- (f1.vwl - mn.f1.grp) / sd.f1.grp
-		  f2.vwl.nrm <- (f2.vwl - mn.f2.grp) / sd.f2.grp
-		  if (reduce) {
-			  f1.vwl.nrm <- mean(f1.vwl.nrm)
-			  f2.vwl.nrm <- mean(f2.vwl.nrm)
-		  }
-		  data.frame(f1=f1.vwl.nrm, f2=f2.vwl.nrm)
-	  })
+    f1.grp <- df.grp[,f1]
+    f2.grp <- df.grp[,f2]
+    mn.f1.grp <- mean(f1.grp, na.rm=TRUE)
+    mn.f2.grp <- mean(f2.grp, na.rm=TRUE)
+    sd.f1.grp <- sd(f1.grp, na.rm=TRUE)
+    sd.f2.grp <- sd(f2.grp, na.rm=TRUE)
+    ddply(df.grp, vowel, function(df.vwl) {
+      f1.vwl <- df.vwl[,f1]
+      f2.vwl <- df.vwl[,f2]
+      f1.vwl.nrm <- (f1.vwl - mn.f1.grp) / sd.f1.grp
+      f2.vwl.nrm <- (f2.vwl - mn.f2.grp) / sd.f2.grp
+      if (reduce) {
+        f1.vwl.nrm <- mean(f1.vwl.nrm)
+        f2.vwl.nrm <- mean(f2.vwl.nrm)
+      }
+      data.frame(f1=f1.vwl.nrm, f2=f2.vwl.nrm)
+    })
   })
 }
 
@@ -36,8 +36,8 @@ ipa <- {}
 map <- "heed /i:/ hid /ɪ/ head /e/ heard /ɜ:/ had /æ/ hud /ʌ/ hard /ɑ:/ hod /ɒ/ hoard /ɔ:/ whod /u:/ hood /ʊ/"
 matches <- as.data.frame(str_match_all(map, "([a-z]+)\\s+/([^/]{1,2})/")[[1]])
 invisible(apply(matches, 1, FUN=function(row){
-	ipa[row["V2"]] <<- row["V3"]
-	monophthongs <<- c(monophthongs, row["V2"])
+  ipa[row["V2"]] <<- row["V3"]
+  monophthongs <<- c(monophthongs, row["V2"])
 }))
 
 
@@ -57,23 +57,30 @@ sse.lob.df <- rbind(
 sse.lob.df$angle <- 0
 sse.lob.df$dist <- 0.25
 
-vwl.df <- rbind(
-	read.csv("vowels-pre-LV.csv"),
-	read.csv("vowels-post-LV.csv"),
-	read.csv("vowels-pre-HV.csv"),
-	read.csv("vowels-post-HV.csv"))
-vwl.df <- vwl.df[vwl.df$vowel %in% monophthongs,]
-vwl.df$speaker <- sapply(vwl.df$participant, FUN=function(x) sub("(C[0-9]+).*?([LH]V)", "\\1\\2", x))
+if (!file.exists('lob.csv')) {
+  cat('Calculating Lobanov formats.')
+  vwl.df <- rbind(
+    read.csv("vowels-pre-LV.csv"),
+    read.csv("vowels-post-LV.csv"),
+    read.csv("vowels-pre-HV.csv"),
+    read.csv("vowels-post-HV.csv"))
+  vwl.df <- vwl.df[vwl.df$vowel %in% monophthongs,]
+  vwl.df$speaker <- sapply(vwl.df$participant, FUN=function(x) sub("(C[0-9]+).*?([LH]V)", "\\1\\2", x))
 
 
 
-lob.df <- lobanov(vwl.df, f1="f1.50", f2="f2.50", group=c("group", "test", "speaker"))
-write.csv(lob.df, "vowels-lob.csv", row.names=FALSE, quote=FALSE)
-
+  lob.df <- lobanov(vwl.df, f1="f1.50", f2="f2.50", group=c("group", "test", "speaker"))
+  write.csv(lob.df, "vowels-lob.csv", row.names=FALSE, quote=FALSE)
+} else {
+  cat('Reading precalculated Lobanov formats\n')
+  lob.df <- read.csv('lob.csv')
+  lob.df$test <- as.factor(lob.df$test)
+  lob.df$test <- relevel(lob.df$test, ref="pre")
+}
 
 
 lob.mn.df <- ddply(lob.df, c("group", "test", "vowel"), function(subset) {
-	data.frame(f1=median(subset$f1), f2=median(subset$f2))
+  data.frame(f1=mean(subset$f1), f2=mean(subset$f2))
 })
 
 # Do plot
@@ -89,8 +96,8 @@ df <- dcast(df.melt, group+vowel~variable+test)
 
 # Roughtly Position labels
 df$angle <- apply(df, 1, FUN=function(d) {
-	atan2(as.numeric(d["f1_post"]) - as.numeric(d["f1_pre"]),
-		  as.numeric(d["f2_post"]) - as.numeric(d["f2_pre"])) / pi * 180 })
+  atan2(as.numeric(d["f1_post"]) - as.numeric(d["f1_pre"]),
+      as.numeric(d["f2_post"]) - as.numeric(d["f2_pre"])) / pi * 180 })
 df$angle <- floor(df$angle / 90) * 90 + 90
 # df$angle[df$vowel == "hud" & df$group == "LV"] <- 180
 # df$angle[df$vowel == "hard" & df$group == "LV"] <- 45
@@ -165,13 +172,19 @@ transforms <- dict({
     "ɜ:" <- 180
     "ʌ" <- 180
     "æ" <- 180
+    "e" <- 180
+    "ɔ:" <- 90
   })
   HV <- dict({
+    "i:" <- 90
     "æ" <- 180
     "ɪ" <- 270
     "ɜ:" <- 180
+    "e" <- 90
     "ɑ:" <- 0
     "ʌ" <- 45
+    "u:" <- 90
+    "ɒ" <- 270
   })
 })
 
@@ -209,73 +222,74 @@ height = 4.5
 
 # Start plot and set some theme stuff
 p <- ggplot(data=) + theme(
-	text=element_text(family="Cabin", size=fontSize),
-	panel.background=element_rect(fill=colors$panel.background),
-	panel.grid.major=element_line(
-		color=colors$panel.grid,
-		linetype="13",
-		lineend="round"),
-	axis.ticks=element_blank(),
-	legend.position="bottom",
-	legend.key=element_rect(
-	  fill="transparent",
-	  colour="transparent"))
+  text=element_text(family="Cabin", size=fontSize),
+  panel.background=element_rect(fill=colors$panel.background),
+  panel.grid.major=element_line(
+    color=colors$panel.grid,
+    linetype="13",
+    lineend="round"),
+  axis.ticks=element_blank(),
+  legend.position="bottom",
+  legend.key=element_rect(
+    fill="transparent",
+    colour="transparent"),
+  axis.title.x = element_text(margin = margin(t = 10, r = 0, b = -10, l = 0)))
 # Reverse scales
 p <- p + scale_y_reverse(
-	expand=c(0.02,0.02),
-	minor_breaks=c(),
-	breaks=seq(-2,2,1),
-	labels=seq(-2,2,1))
+  expand=c(0.02,0.02),
+  minor_breaks=c(),
+  breaks=seq(-2,2,1),
+  labels=seq(-2,2,1))
 p <- p + scale_x_reverse(
-	expand=c(0.02,0.02),
-	minor_breaks=c(),
-	breaks=seq(-2,2,1),
-	labels=seq(-2,2,1))
+  expand=c(0.02,0.02),
+  minor_breaks=c(),
+  breaks=seq(-2,2,1),
+  labels=seq(-2,2,1))
 # Add limits here to prevent filtering of data
 p <- p + coord_cartesian(xlim=c(2,-2), ylim=c(2, -2))
 
 # Add SSBE points...
 p <- p + geom_point(
-	data=sse.lob.df,
-	aes(x=f2, y=f1, fill="SSBE"),
+  data=sse.lob.df,
+  aes(x=f2, y=f1, fill="SSBE"),
   color=colors$sse,
-	size=6.25)
+  size=6.25)
 # ...and SSBE labels
 p <- p + geom_text(
-	data=sse.lob.df,
-	aes(x=f2, y=f1, label=ipa),
+  data=sse.lob.df,
+  aes(x=f2, y=f1, label=ipa),
   color=colors$sse.label,
-	family="DejaVuSans",
+  family="DejaVuSans",
   vjust=0.4,
-	size=fontSize*0.333)
+  size=fontSize*0.333)
 # Add an extra legend entry
 p <- p + scale_fill_manual("", breaks = "SSBE", values=colors$sse)
 
 # Draw lines between pre and post
 p <- p + geom_segment(
-	data=df,
-	aes(x=f2_pre, xend=f2_post, y=f1_pre, yend=f1_post),
-	# arrow=arrow(type="closed", length=unit(0.075, "in")),
-	inherit.aes=FALSE,
-	colour=colors$arrow)
+  data=df,
+  aes(x=f2_pre, xend=f2_post, y=f1_pre, yend=f1_post),
+  # arrow=arrow(type="closed", length=unit(0.075, "in")),
+  inherit.aes=FALSE,
+  colour=colors$arrow)
 
 # Draw pre and post points
 p <- p + geom_point(
-	data=lob.mn.df,
-	aes(x=f2, y=f1, color=test),
-	size=3)
+  data=lob.mn.df,
+  aes(x=f2, y=f1, color=test),
+  size=3)
 p <- p + scale_color_manual(
   values=c(colors$pre, colors$post),
   name="Test")
 
 # Add labels to pre points
 p <- p + geom_text(
-	data=lab.df,
-	aes(x, y, label=label),
-	colour=colors$ipa,
-	family="DejaVuSans",
+  data=lab.df,
+  aes(x, y, label=label),
+  colour=colors$ipa,
+  family="DejaVuSans",
   vjust=0.4,
-	size=fontSize*0.4)
+  size=fontSize*0.4)
 
 p <- p +  ylab("F1 (Lobanov)")
 p <- p +  xlab("F2 (Lobanov)")
